refactor: move tests to correct directory and adjust
